READ DESCRIPTION
MAKE SURE TO CHANGE THE POSITION OF THE WEIGHTED COLUMN IN THE CODE. It will always hide the X'th position in the gradebook after screenreader is turned on. So right now it is set to 5, it will hide Weighted Total, then it will hide Total (since Total cascades down to be the 5th column afterwards).

SMARTVIEWS: Verify the order of the Category codes to make sure the script will accurately grab the right category code. (remember Assignments will always be 0, Discussions is generally 3, and Tests will be 9)